// types/type.ts
export interface ICategoryType {
    name: string;
    image: string;
    description: string;
    rating: number;
    price: string;
    duration: string;
    location: string;
    category: string;
    id: string;
}

export interface IGroupsType {
    id: number,
    name: string;
    image: string;
    rating: number;
    reviews: number;
}